feat: event取得時に、デザインに必要な参加ユーザ情報も取得するように変更 既存のJoinedUserInfoを使用したが、必要でない項目も送信しているので、変えてもいいかもしれない
