+++ springboot/src/main/java/com/asakatu/response/ForFrontEvent.java
@@ -1,12 +1,16 @@
 package com.asakatu.response;
 
+import com.asakatu.controller.JoinedUserInfo;
 import com.asakatu.entity.Event;
+
+import java.util.List;
 
 // フロントのための整形されたEventClassです。追加したい情報があったらこちらへ
 public class ForFrontEvent {
     private String designDate;
     private Event event;
     private boolean hasJoin;
+    private List<JoinedUserInfo> joinedUserInfo;
 
     public String getDesignDate() {
         return designDate;
@@ -30,5 +34,13 @@ public class ForFrontEvent {
 
     public void setHasJoin(boolean hasJoin) {
         this.hasJoin = hasJoin;
+    }
+
+    public List<JoinedUserInfo> getJoinedUserInfo() {
+        return joinedUserInfo;
+    }
+
+    public void setJoinedUserInfo(List<JoinedUserInfo> joinedUserInfo) {
+        this.joinedUserInfo = joinedUserInfo;
     }
 }
