+++ springboot/src/main/java/com/asakatu/service/PostService.java
@@ -1,22 +1,20 @@
 package com.asakatu.service;
 
+import com.asakatu.controller.JoinedUserInfo;
 import com.asakatu.entity.Event;
 import com.asakatu.entity.User;
-import com.asakatu.entity.UserStatus;
 import com.asakatu.repository.UserRepository;
 import com.asakatu.response.ForFrontEvent;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.TextStyle;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
-import java.util.Optional;
 
 @Service
 public class PostService{
@@ -59,6 +57,16 @@ public class PostService{
         List<User> eventJoinedUsersList = userRepository.findUsersByEventsListIn(event);
 
         forFrontEvent.setHasJoin(eventJoinedUsersList.contains(user));
+
+        List<JoinedUserInfo> joinedUserInfoList = new ArrayList<>();
+        for (User joinedUser : eventJoinedUsersList) {
+            JoinedUserInfo joinedUserInfo = new JoinedUserInfo();
+            joinedUserInfo.setImagePath(joinedUser.getImagePath());
+            joinedUserInfo.setId(joinedUser.getId());
+            joinedUserInfoList.add(joinedUserInfo);
+        }
+        forFrontEvent.setJoinedUserInfo(joinedUserInfoList);
+
         return forFrontEvent;
     }
 }
